pyproject.toml
[project]
name = "instante-infinito"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "pyjwt>=2.10.1",
    "pytelegrambotapi>=4.27.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.2",
]
